숫자 자릿수 지정 및 타입 변경

+++ backend/src/main/java/com/eyes/solstory/domain/story/entity/StoryCard.java
@@ -17,8 +17,8 @@ public class StoryCard {
 	// 카드 일련번호
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "card_no")
-    private Long cardNo;
+    @Column(name = "card_no", precision = 10)
+    private int cardNo;
 
     // 카드 종류(인물/아이템/상황)
     @Column(name = "card_type", nullable = false, length = 50)
@@ -29,6 +29,6 @@ public class StoryCard {
     private String cardKeyword;
 
     // 카드를 얻기 위해 필요한 열쇠 수
-    @Column(name = "keys_required", nullable = false)
-    private Integer keysRequired;
+    @Column(name = "keys_required", nullable = false, precision = 2)
+    private int keysRequired;
 }
